requirements.txt
numpy
PyOpenGL
pynput
playsound
pysdl2
pyopengltk